refactor stuctrure files
