+++ src/main/java/org/fundacionjala/converter/controller/AudioController.java
@@ -5,11 +5,11 @@ import org.fundacionjala.converter.executor.Executor;
 import org.fundacionjala.converter.model.ChecksumMD5;
 import org.fundacionjala.converter.model.command.AudioModel;
 import org.fundacionjala.converter.model.command.ICommand;
-import org.fundacionjala.converter.model.entity.File;
+import org.fundacionjala.converter.database.entity.File;
 import org.fundacionjala.converter.model.parameter.ModelParameter;
 import org.fundacionjala.converter.model.parameter.multimedia.AudioParameter;
-import org.fundacionjala.converter.model.service.FileService;
-import org.fundacionjala.converter.model.service.FileUploadService;
+import org.fundacionjala.converter.controller.service.FileService;
+import org.fundacionjala.converter.controller.service.FileUploadService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.RequestMapping;
