+++ src/main/java/org/fundacionjala/converter/controller/ExtractTextController.java
@@ -9,8 +9,8 @@
 package org.fundacionjala.converter.controller;
 
 import org.fundacionjala.converter.controller.request.RequestExtractTextParameter;
-import org.fundacionjala.converter.model.entity.File;
-import org.fundacionjala.converter.model.service.FileService;
+import org.fundacionjala.converter.database.entity.File;
+import org.fundacionjala.converter.controller.service.FileService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.RequestMapping;
