+++ src/main/java/org/fundacionjala/converter/controller/ImageController.java
@@ -12,8 +12,8 @@ import org.fundacionjala.converter.controller.request.RequestImageParameter;
 import org.fundacionjala.converter.executor.Executor;
 import org.fundacionjala.converter.model.command.ImageModel;
 import org.fundacionjala.converter.model.parameter.image.ImageParameter;
-import org.fundacionjala.converter.model.service.FileService;
-import org.fundacionjala.converter.model.service.FileUploadService;
+import org.fundacionjala.converter.controller.service.FileService;
+import org.fundacionjala.converter.controller.service.FileUploadService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.RestController;
