+++ src/main/java/org/fundacionjala/converter/controller/MetadataController.java
@@ -7,14 +7,14 @@
  * license agreement you entered into with Fundacion Jala
  */
 package org.fundacionjala.converter.controller;
-import org.fundacionjala.converter.model.entity.File;
+import org.fundacionjala.converter.database.entity.File;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
-import org.fundacionjala.converter.model.service.FileService;
+import org.fundacionjala.converter.controller.service.FileService;
 import org.fundacionjala.converter.controller.request.RequestMetadataParameter;
 
 import java.nio.file.Files;
